Add file with hashing functions for tests
