+++ src/tests/test_ucc_algorithms.cpp
@@ -14,6 +14,7 @@
 #include "config/names.h"
 #include "config/thread_number/type.h"
 #include "table_config.h"
+#include "test_hash_util.h"
 
 std::ostream& operator<<(std::ostream& os, Vertical const& v) {
     os << v.ToString();
@@ -23,29 +24,6 @@ std::ostream& operator<<(std::ostream& os, Vertical const& v) {
 namespace tests {
 
 namespace {
-
-// Implement custom hash functions since implementation of `std::hash` or `boost::hash` may change
-// depending on the library version/architecture/os/whatever leading to tests failing.
-// Taken from
-// https://stackoverflow.com/questions/20511347/a-good-hash-function-for-a-vector/72073933#72073933
-std::size_t Hash(std::vector<unsigned> const& vec) {
-    std::size_t seed = vec.size();
-    for (auto x : vec) {
-        x = ((x >> 16) ^ x) * 0x45d9f3b;
-        x = ((x >> 16) ^ x) * 0x45d9f3b;
-        x = (x >> 16) ^ x;
-        seed ^= x + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-    }
-    return seed;
-}
-
-std::size_t Hash(std::vector<std::vector<unsigned>> const& vec) {
-    size_t hash = 1;
-    for (auto const& v : vec) {
-        hash = 31 * hash + Hash(v);
-    }
-    return hash;
-}
 
 // TODO(polyntsov): think how we should organize test code, maybe implement some classes with
 // basic functionality common for all primitive mining algorithms. Without it every test class
